requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
executing==0.8.3
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
msgpack==1.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
proto-plus==1.20.6
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
